Put my tests in newly created class CddUnitTest, as to leave CddTest as it was
